Minor fix to the hashing since it wasn't working when i tested it. System looks like it's back to fully working and seeing brains now

+++ simulator/brain.h
@@ -20,7 +20,7 @@ typedef struct brain
  float multL2[AG_CONNS_L2];
  float mutationRate;
  unsigned char latestDecision;
- long long speciesHash;
+ unsigned long speciesHash;
  unsigned int brainSize;
 } brain;
